More bug fixes. Good practices are for noobs.

+++ src/execution/set_fds.c
@@ -6,7 +6,7 @@
 /*   By: ccamargo <ccamargo@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/02 17:16:09 by ccamargo          #+#    #+#             */
-/*   Updated: 2023/05/04 17:33:43 by ccamargo         ###   ########.fr       */
+/*   Updated: 2023/05/11 20:46:13 by ccamargo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void	set_io_redirection(t_tk_exec *exec_vars, int cmd_n)
 
 void	set_fd_variables(t_tk_exec *exec_vars)
 {
-	if (exec_vars->fd_in == exec_vars->fd_heredoc)
+	if (exec_vars->fd_in == exec_vars->fd_heredoc && exec_vars->fd_in != -1)
 	{
 		if (unlink(".tmp") != 0)
 			perror("unlink() failed");
